+++ utils/index.ts
@@ -1,4 +1,4 @@
 export { dummyImages } from 'utils/dummyImages';
 export { calculatingDDay } from 'utils/calculatingDDay';
 export { formatDate } from 'utils/formatDate';
-export { default as FormValidation } from 'utils/formValidation';
+export { signupValidation, postValidation } from 'utils/formValidation';
